backend/requirements.txt
# FastAPI and ASGI
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# ML and Image Processing
tensorflow==2.15.0
opencv-python==4.8.1.78
pillow==10.1.0
numpy==1.24.3
scikit-learn==1.3.2

# Text-to-Speech
gtts==2.4.0
pyttsx3==2.90

# WhatsApp Integration
twilio==8.10.3
requests==2.31.0

# Weather API
pyowm==3.3.0

# Utilities
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
httpx==0.25.2
redis==5.0.1
celery==5.3.4

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0
mypy==1.7.1
